loginUser-logoutUser-createUser

+++ myproject/base/views.py
@@ -1,8 +1,62 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from . models import Room , Topic
 from .forms import RoomForm
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib import messages
+from django.contrib.auth import authenticate , login , logout
 
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username = username)
+        except:
+            messages.error(request,'User does not exist')
+
+        user = authenticate(request , username = username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request,'Username OR password does not exit')
+
+    context = {
+        'page':page
+    }
+    return render(request, 'login_register.html',context)
+
+def logoutPage(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured during registration')
+    context = {
+        'form':form
+    }
+    return render(request,'login_register.html',context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -24,8 +78,14 @@ def room(request,pk):
 
     return render(request,'room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
+    
+    if request.user != None :
+        messages.error(request,'you must login')
+        return redirect('home')
+    
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -35,9 +95,16 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host :
+        messages.warning(request,'Your are not allowed here!!')
+        return redirect('home')
+
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -47,8 +114,14 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request, 'room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host :
+        messages.warning(request,'Your are not allowed here!!')
+        return redirect('home')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
